Committing changes before redoing contract sync cause I'm stupid...whoops
So contracts are generated dynamically...duh. That means i cant simply
save the contract title to the database and then set the state of that
specific contract from the list (that doesn't exist) when syncing. So I
need to somehow save all of the data from a contract to the database
(Serialize?) and then rebuild/regenerate that contract to all team
members on sync